+++ python/local_path_visualizer.py
@@ -5,6 +5,7 @@ import math
 from std_msgs.msg import Float64
 from local_pathfinding.msg import path, latlon
 import utilities as utils
+import obstacles as obs
 import Sailbot as sbot
 from matplotlib import pyplot as plt
 from matplotlib import patches
@@ -179,7 +180,7 @@ if __name__ == '__main__':
         localPathXY = [utils.latlonToXY(localWaypoint, referenceLatlon) for localWaypoint in localPath]
         localPathX = [xy[0] for xy in localPathXY]
         localPathY = [xy[1] for xy in localPathXY]
-        shipsXY = utils.getObstacles(state, referenceLatlon)
+        shipsXY = obs.getObstacles(state, referenceLatlon)
 
         # Update plots
         localPathPlot.set_xdata(localPathX)
