+++ test/test_utilities.py
@@ -7,6 +7,8 @@ from local_pathfinding.msg import latlon, AISMsg, AISShip
 import Sailbot as sbot
 import rostest
 import unittest
+import matplotlib.pyplot as plt
+
 
 # Do something with local_imports to avoid lint errors
 local_imports.printMessage()
@@ -87,60 +89,6 @@ class TestUtilities(unittest.TestCase):
                            .destination(point=(position.lat, position.lon), bearing=utils.BEARING_EAST))
         reachedGlobalWaypoint = latlon(reachedPosition.latitude, reachedPosition.longitude)
         self.assertTrue(utils.globalWaypointReached(position, reachedGlobalWaypoint))
-
-    def test_getObstacles(self):
-        # Setup: sailbot starting at (0,0). obstacle at (1,0) heading west
-        currentLatlon = latlon(0, 0)
-        referenceLatlon = currentLatlon
-        shipLatlon = utils.XYToLatlon(xy=(1, 0), referenceLatlon=referenceLatlon)
-        ships = [AISShip(ID=1000, lat=shipLatlon.lat, lon=shipLatlon.lon, headingDegrees=utils.HEADING_WEST,
-                         speedKmph=1)]
-        obstacles = utils.getObstacles(ships=ships, position=currentLatlon, speedKmph=1,
-                                       referenceLatlon=referenceLatlon)
-
-        self.assertFalse(utils.isValid(xy=[0, 0], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[1, 0], obstacles=obstacles))
-        self.assertTrue(utils.isValid(xy=[2, 0], obstacles=obstacles))
-
-    def test_getObstacles2(self):
-        # Setup starting at (0,0) with obstacle at (1,1) heading south-west
-        currentLatlon = latlon(0, 0)
-        referenceLatlon = currentLatlon
-        shipLatlon = utils.XYToLatlon(xy=(1, 1), referenceLatlon=referenceLatlon)
-        ships = [AISShip(ID=1000, lat=shipLatlon.lat, lon=shipLatlon.lon, headingDegrees=225, speedKmph=2**0.5)]
-        obstacles = utils.getObstacles(ships=ships, position=currentLatlon, speedKmph=1,
-                                       referenceLatlon=referenceLatlon)
-# Uncomment below to see obstacles extended on a plot
-#        ax = plt.gca()
-#        for obstacle in obstacles:
-#            ax.add_patch(plt.Circle((obstacle.x, obstacle.y), radius=obstacle.radius))
-#        plt.show()
-        self.assertFalse(utils.isValid(xy=[1, 1], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[0, 0], obstacles=obstacles))
-        self.assertTrue(utils.isValid(xy=[-1, -1], obstacles=obstacles))
-
-    def test_getObstacles3(self):
-        # Setup starting at (0,0) with obstacles at (0,3) and (-1,-1)
-        currentLatlon = latlon(0, 0)
-        referenceLatlon = currentLatlon
-        shipLatlon = utils.XYToLatlon(xy=(0, 3), referenceLatlon=referenceLatlon)
-        shipLatlon2 = utils.XYToLatlon(xy=(-1, -1), referenceLatlon=referenceLatlon)
-        ship1 = AISShip(ID=1000, lat=shipLatlon.lat, lon=shipLatlon.lon, headingDegrees=270, speedKmph=1.5)
-        ship2 = AISShip(ID=1001, lat=shipLatlon2.lat, lon=shipLatlon2.lon, headingDegrees=45, speedKmph=10)
-        obstacles = utils.getObstacles(ships=[ship1, ship2], position=currentLatlon,
-                                       speedKmph=1, referenceLatlon=referenceLatlon)
-# Uncomment below to see obstacles extended on a plot
-#        ax = plt.gca()
-#        for obstacle in obstacles:
-#            ax.add_patch(plt.Circle((obstacle.x, obstacle.y), radius=obstacle.radius))
-#        plt.show()
-        self.assertFalse(utils.isValid(xy=[0, 0], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[1, 1], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[3, 3], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[0, -1], obstacles=obstacles))
-        self.assertTrue(utils.isValid(xy=[0, 4], obstacles=obstacles))
-        self.assertFalse(utils.isValid(xy=[0, -1.19], obstacles=obstacles))
-        self.assertTrue(utils.isValid(xy=[0, -2.3], obstacles=obstacles))
 
     def test_headingToBearingDegrees(self):
         # Basic tests
